requirements.txt
Flask==3.1.0
gunicorn
Werkzeug==3.1.3
pyodbc==5.2.0
pandas==2.2.3
azure-identity==1.20.0
azure-storage-blob==12.24.0
